# GameWebApplication/src/main/resources/com/tcoffman/ttwb/web/resource/swagger.json
{"swagger":"2.0","info":{"description":"Access to view and update game states","version":"1.0.0","title":""},"host":"localhost:8080","basePath":"/tabletop-web/api","tags":[{"name":"Models"},{"name":"States"}],"schemes":["http"],"paths":{"/models":{"get":{"tags":["Models"],"summary":"List of all models","description":"","operationId":"getModels","produces":["application/xml","application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/ModelResource"}}}}}},"/models/{modelId}/plugins":{"get":{"tags":["Models"],"operationId":"getPlugins","produces":["application/xml","application/json"],"parameters":[{"name":"modelId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/ModelPluginResource"}},"headers":{}}}}},"/models/{modelId}/parts":{"get":{"tags":["Models"],"operationId":"getParts","produces":["application/xml","application/json"],"parameters":[{"name":"modelId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/PartInstanceResource"}},"headers":{}}}}},"/models/{modelId}/roles":{"get":{"tags":["Models"],"operationId":"getRoles","produces":["application/xml","application/json"],"parameters":[{"name":"modelId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/RoleResource"}},"headers":{}}}}},"/models/{modelId}/prototypes":{"get":{"tags":["Models"],"operationId":"getPartPrototypes","produces":["application/xml","application/json"],"parameters":[{"name":"modelId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/PartPrototypeResource"}},"headers":{}}}}},"/models/{modelId}/prototypes/{prototypeId}/places":{"get":{"tags":["Models"],"operationId":"getPlacePrototypes","produces":["application/xml","application/json"],"parameters":[{"name":"prototypeId","in":"path","required":true,"type":"string"},{"name":"modelId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/PlacePrototypeResource"}},"headers":{}}}}},"/models/{modelId}/stages/{stageId}/rules":{"get":{"tags":["Models"],"operationId":"getRules","produces":["application/xml","application/json"],"parameters":[{"name":"stageId","in":"path","required":true,"type":"string"},{"name":"modelId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/RuleResource"}},"headers":{}}}}},"/models/{modelId}/stages":{"get":{"tags":["Models"],"operationId":"getStages","produces":["application/xml","application/json"],"parameters":[{"name":"modelId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/StageResource"}},"headers":{}}}}},"/states/{stateId}":{"get":{"tags":["States"],"operationId":"get","produces":["application/xml","application/json"],"parameters":[{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/StateResource"},"headers":{}}}},"delete":{"tags":["States"],"operationId":"delete","produces":["application/xml","application/json"],"parameters":[{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/states/{stateId}/find":{"post":{"tags":["States"],"operationId":"find","consumes":["application/json"],"produces":["application/xml","application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/PlacePatternForm"}},{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/PlaceResource"}},"headers":{}}}}},"/states/{stateId}/log/{logEntryIndex}":{"get":{"tags":["States"],"operationId":"get","produces":["application/xml","application/json"],"parameters":[{"name":"logEntryIndex","in":"path","required":true,"type":"integer","format":"int64"},{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/LogEntryResource"},"headers":{}}}}},"/states/{stateId}/log":{"get":{"tags":["States"],"operationId":"getLogEntries","produces":["application/xml","application/json"],"parameters":[{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/LogEntryResource"}},"headers":{}}}},"post":{"tags":["States"],"operationId":"mutate","consumes":["application/json"],"produces":["application/xml","application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/StateMutationForm"}},{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/LogEntryResource"},"headers":{}}}}},"/states/{stateId}/participants":{"get":{"tags":["States"],"operationId":"getParticipants","produces":["application/xml","application/json"],"parameters":[{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/ParticipantResource"}},"headers":{}}}},"post":{"tags":["States"],"operationId":"createParticipant","consumes":["application/json"],"produces":["application/xml","application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/ParticipantCreationForm"}},{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/ParticipantResource"},"headers":{}}}}},"/states/{stateId}/relationships":{"get":{"tags":["States"],"operationId":"getRelationships","produces":["application/xml","application/json"],"parameters":[{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/RelationshipResource"}},"headers":{}}}}},"/states/{stateId}/allowedOperations":{"get":{"tags":["States"],"operationId":"getOperationPatternSets","produces":["application/xml","application/json"],"parameters":[{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/OperationPatternSetResource"}},"headers":{}}}}},"/states/{stateId}/parts":{"get":{"tags":["States"],"operationId":"getParts","produces":["application/xml","application/json"],"parameters":[{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/PartResource"}},"headers":{}}}}},"/states/{stateId}/parts/{partId}/places/{pluginUri}/{placeTypeLocalName}":{"get":{"tags":["States"],"operationId":"get","produces":["application/xml","application/json"],"parameters":[{"name":"pluginUri","in":"path","required":true,"type":"string"},{"name":"placeTypeLocalName","in":"path","required":true,"type":"string"},{"name":"partId","in":"path","required":true,"type":"string"},{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/PlaceResource"},"headers":{}}}}},"/states/{stateId}/parts/{partId}/places":{"get":{"tags":["States"],"operationId":"getPlaces","produces":["application/xml","application/json"],"parameters":[{"name":"partId","in":"path","required":true,"type":"string"},{"name":"stateId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/PlaceResource"}},"headers":{}}}}},"/states":{"get":{"tags":["States"],"summary":"List of all states","description":"","operationId":"getStates","produces":["application/xml","application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"type":"string","format":"uri"}}}}},"post":{"tags":["States"],"operationId":"createState","consumes":["application/json"],"produces":["application/xml","application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/StateCreationForm"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/StateResource"},"headers":{}}}}}},"definitions":{"ModelPluginResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"relationshipTypes":{"$ref":"#/definitions/RelationshipTypesResource"},"label":{"type":"string"},"placeTypes":{"$ref":"#/definitions/PlaceTypesResource"}}},"ModelPluginsResource":{"type":"object","properties":{"plugins":{"type":"array","items":{"$ref":"#/definitions/ModelPluginResource"}}}},"ModelResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"plugins":{"$ref":"#/definitions/ModelPluginsResource"},"parts":{"$ref":"#/definitions/PartInstancesResource"},"roles":{"$ref":"#/definitions/RolesResource"},"partPrototypes":{"$ref":"#/definitions/PartPrototypesResource"},"label":{"type":"string"},"stages":{"$ref":"#/definitions/StagesResource"}}},"OperationPatternResource":{"type":"object","properties":{"type":{"type":"string"},"target":{"$ref":"#/definitions/PlacePatternResourceObject"},"subject":{"$ref":"#/definitions/PlacePatternResourceObject"},"quantity":{"$ref":"#/definitions/QuantityPatternResource"}}},"PartInstanceResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"},"prototypeResource":{"type":"string","format":"uri"}}},"PartInstancesResource":{"type":"object","properties":{"parts":{"type":"array","items":{"$ref":"#/definitions/PartInstanceResource"}}}},"PartPatternResource":{"type":"object","properties":{"type":{"type":"string"}}},"PartPatternResourceObject":{"type":"object","properties":{"type":{"type":"string"}}},"PartPrototypeResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"},"places":{"$ref":"#/definitions/PlacePrototypesResource"}}},"PartPrototypesResource":{"type":"object","properties":{"partPrototypes":{"type":"array","items":{"$ref":"#/definitions/PartPrototypeResource"}}}},"PlacePatternResource":{"type":"object","properties":{"type":{"type":"string"},"part":{"$ref":"#/definitions/PartPatternResourceObject"},"label":{"type":"string"}}},"PlacePatternResourceObject":{"type":"object","properties":{"type":{"type":"string"},"part":{"$ref":"#/definitions/PartPatternResourceObject"},"label":{"type":"string"}}},"PlacePrototypeResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{"type":"string"}},"placeType":{"type":"string"},"label":{"type":"string"}}},"PlacePrototypesResource":{"type":"object","properties":{"placePrototypes":{"type":"array","items":{"$ref":"#/definitions/PlacePrototypeResource"}}}},"PlaceTypesResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"}}},"QuantityPatternResource":{"type":"object","properties":{"type":{"type":"string"},"label":{"type":"string"},"maximum":{"type":"integer","format":"int64"},"minimum":{"type":"integer","format":"int64"}}},"RelationshipTypesResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"}}},"RoleResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"}}},"RolesResource":{"type":"object","properties":{"roles":{"type":"array","items":{"$ref":"#/definitions/RoleResource"}}}},"RuleResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"},"resultLabel":{"type":"string"},"resultResource":{"type":"string","format":"uri"},"operationPatterns":{"type":"array","items":{"$ref":"#/definitions/OperationPatternResource"}}}},"RulesResource":{"type":"object","properties":{"rules":{"type":"array","items":{"$ref":"#/definitions/RuleResource"}}}},"StageResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"rules":{"$ref":"#/definitions/RulesResource"},"label":{"type":"string"},"stages":{"$ref":"#/definitions/StagesResource"}}},"StagesResource":{"type":"object","properties":{"stages":{"type":"array","items":{"$ref":"#/definitions/StageResource"}}}},"RelationshipTypeResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"}}},"PlaceTypeResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"}}},"LogEntriesResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"logEntries":{"type":"array","items":{"$ref":"#/definitions/LogEntryResource"}}}},"LogEntryResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"rollback":{"type":"string","format":"uri"},"rollbackMutations":{"type":"array","items":{"$ref":"#/definitions/StateMutationResource"}},"forwardMutations":{"type":"array","items":{"$ref":"#/definitions/StateMutationResource"}},"forward":{"type":"string","format":"uri"},"label":{"type":"string"}}},"OperationPatternSetResource":{"type":"object","properties":{"label":{"type":"string"},"resultLabel":{"type":"string"},"resultResource":{"type":"string","format":"uri"},"operationPatterns":{"type":"array","items":{"$ref":"#/definitions/OperationPatternResource"}}}},"OperationPatternSetsResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"operationPatternSets":{"type":"array","items":{"$ref":"#/definitions/OperationPatternSetResource"}}}},"PartResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"},"prototypeResource":{"type":"string","format":"uri"},"places":{"$ref":"#/definitions/PlacesResource"}}},"ParticipantResource":{"type":"object","properties":{"label":{"type":"string"},"roleResource":{"type":"string","format":"uri"}}},"ParticipantsResource":{"type":"object","properties":{"participants":{"type":"array","items":{"$ref":"#/definitions/ParticipantResource"}}}},"PartsResource":{"type":"object","properties":{"parts":{"type":"array","items":{"$ref":"#/definitions/PartResource"}}}},"PlaceResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"label":{"type":"string"},"typeResource":{"type":"string","format":"uri"}}},"PlacesResource":{"type":"object","properties":{"places":{"type":"array","items":{"$ref":"#/definitions/PlaceResource"}}}},"RelationshipResource":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"source":{"type":"string"},"destination":{"type":"string"},"label":{"type":"string"}}},"RelationshipsResource":{"type":"object","properties":{"relationships":{"type":"array","items":{"$ref":"#/definitions/RelationshipResource"}}}},"StateMutationResource":{"type":"object","properties":{"source":{"type":"string","format":"uri"},"relationshipType":{"type":"string","format":"uri"},"role":{"type":"string","format":"uri"},"operationType":{"type":"string","enum":["SIGNAL","MOVE","ORIENT","JOIN","SPLIT"]},"mutation":{"type":"string"},"destination":{"type":"string","format":"uri"},"label":{"type":"string"}}},"StateResource":{"type":"object","properties":{"resource":{"type":"string","format":"uri"},"log":{"$ref":"#/definitions/LogEntriesResource"},"model":{"type":"string","format":"uri"},"participants":{"$ref":"#/definitions/ParticipantsResource"},"relationships":{"$ref":"#/definitions/RelationshipsResource"},"allowedOperations":{"$ref":"#/definitions/OperationPatternSetsResource"},"parts":{"$ref":"#/definitions/PartsResource"},"stage":{"type":"string","format":"uri"},"label":{"type":"string"}}},"PartPatternForm":{"type":"object","properties":{"type":{"type":"string"},"quantity":{"$ref":"#/definitions/QuantityPatternForm"},"pattern":{"$ref":"#/definitions/PartPatternForm"},"patterns":{"type":"array","items":{"$ref":"#/definitions/PartPatternForm"}},"prototype":{"type":"string"},"role":{"type":"string"}}},"PlacePatternForm":{"type":"object","properties":{"type":{"type":"string"},"quantity":{"$ref":"#/definitions/QuantityPatternForm"},"forward":{"type":"boolean"},"pattern":{"$ref":"#/definitions/PlacePatternForm"},"exists":{"type":"boolean"},"part":{"$ref":"#/definitions/PartPatternForm"},"patterns":{"type":"array","items":{"$ref":"#/definitions/PlacePatternForm"}},"placeType":{"type":"string"},"relationshipType":{"type":"string"},"related":{"$ref":"#/definitions/PlacePatternForm"},"role":{"type":"string"}}},"QuantityPatternForm":{"type":"object","properties":{"type":{"type":"string"},"max":{"type":"integer","format":"int32"},"min":{"type":"integer","format":"int32"}}},"StateMutationForm":{"type":"object","properties":{"result":{"type":"string"},"operations":{"type":"array","items":{"$ref":"#/definitions/StateOperationForm"}},"role":{"type":"string"}}},"StateOperationForm":{"type":"object","properties":{"type":{"type":"string"},"target":{"$ref":"#/definitions/PlacePatternForm"},"subject":{"$ref":"#/definitions/PlacePatternForm"}}},"ParticipantCreationForm":{"type":"object","properties":{"user":{"type":"string"},"role":{"type":"string"}}},"StateCreationForm":{"type":"object","properties":{"model":{"type":"string"}}}}}